moved forms to one place in project structure, further form validation
